refactor(shopware): Moved persistense directory into infrastructure
