+++ module/exporter-shopware-6/src/Infrastructure/Persistence/Query/DbalShopware6LanguageQuery.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ExporterShopware6\Persistence\Dbal\Query;
+namespace Ergonode\ExporterShopware6\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Types\Types;
