+++ module/exporter-shopware-6/src/Infrastructure/Persistence/Query/DbalShopware6MultimediaQuery.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -6,25 +7,24 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ExporterShopware6\Persistence\Dbal\Query;
+namespace Ergonode\ExporterShopware6\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Types\Types;
-use Ergonode\ExporterShopware6\Domain\Query\Shopware6PropertyGroupQueryInterface;
-use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
+use Ergonode\ExporterShopware6\Domain\Query\Shopware6MultimediaQueryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\ChannelId;
+use Ergonode\SharedKernel\Domain\Aggregate\MultimediaId;
 
 /**
  */
-class DbalShopware6PropertyGroupQuery implements Shopware6PropertyGroupQueryInterface
+class DbalShopware6MultimediaQuery implements Shopware6MultimediaQueryInterface
 {
-    private const TABLE = 'exporter.shopware6_property_group';
+    private const TABLE = 'exporter.shopware6_multimedia';
     private const FIELDS = [
         'channel_id',
-        'attribute_id',
+        'multimedia_id',
         'shopware6_id',
     ];
-
     /**
      * @var Connection
      */
@@ -39,46 +39,43 @@ class DbalShopware6PropertyGroupQuery implements Shopware6PropertyGroupQueryInte
     }
 
     /**
-     * @param ChannelId $channelId
+     * @param ChannelId $channel
      * @param string    $shopwareId
      *
-     * @return AttributeId|null
+     * @return MultimediaId|null
      */
-    public function loadByShopwareId(ChannelId $channelId, string $shopwareId): ?AttributeId
+    public function loadByShopwareId(ChannelId $channel, string $shopwareId): ?MultimediaId
     {
         $query = $this->connection->createQueryBuilder();
         $record = $query
             ->select(self::FIELDS)
-            ->from(self::TABLE, 'pg')
+            ->from(self::TABLE, 'm')
             ->where($query->expr()->eq('channel_id', ':channelId'))
-            ->setParameter(':channelId', $channelId->getValue())
+            ->setParameter(':channelId', $channel->getValue())
             ->andWhere($query->expr()->eq('shopware6_id', ':shopware6Id'))
             ->setParameter(':shopware6Id', $shopwareId)
             ->execute()
             ->fetch();
 
         if ($record) {
-            return new AttributeId($record['attribute_id']);
+            return new MultimediaId($record['multimedia_id']);
         }
 
         return null;
     }
 
     /**
-     * @param ChannelId          $channelId
+     * @param ChannelId          $channel
      * @param \DateTimeImmutable $dateTime
-     * @param string             $type
      */
-    public function cleanData(ChannelId $channelId, \DateTimeImmutable $dateTime, string $type): void
+    public function cleanData(ChannelId $channel, \DateTimeImmutable $dateTime): void
     {
         $query = $this->connection->createQueryBuilder();
-        $query->delete(self::TABLE, 'pg')
-            ->where($query->expr()->eq('channel_id', ':channelId'))
-            ->setParameter(':channelId', $channelId->getValue())
-            ->andWhere($query->expr()->lt('pg.update_at', ':updateAt'))
+        $query->delete(self::TABLE, 'm')
+            ->where($query->expr()->eq('m.channel_id', ':channelId'))
+            ->setParameter(':channelId', $channel->getValue())
+            ->andWhere($query->expr()->lt('m.update_at', ':updateAt'))
             ->setParameter(':updateAt', $dateTime, Types::DATETIMETZ_MUTABLE)
-            ->andWhere($query->expr()->eq('pg.type', ':type'))
-            ->setParameter(':type', $type)
             ->execute();
     }
 }
