+++ module/exporter-shopware-6/src/Infrastructure/Persistence/Query/DbalShopware6PropertyGroupQuery.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -6,24 +7,25 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ExporterShopware6\Persistence\Dbal\Query;
+namespace Ergonode\ExporterShopware6\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Types\Types;
-use Ergonode\ExporterShopware6\Domain\Query\Shopware6MultimediaQueryInterface;
+use Ergonode\ExporterShopware6\Domain\Query\Shopware6PropertyGroupQueryInterface;
+use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\SharedKernel\Domain\Aggregate\ChannelId;
-use Ergonode\SharedKernel\Domain\Aggregate\MultimediaId;
 
 /**
  */
-class DbalShopware6MultimediaQuery implements Shopware6MultimediaQueryInterface
+class DbalShopware6PropertyGroupQuery implements Shopware6PropertyGroupQueryInterface
 {
-    private const TABLE = 'exporter.shopware6_multimedia';
+    private const TABLE = 'exporter.shopware6_property_group';
     private const FIELDS = [
         'channel_id',
-        'multimedia_id',
+        'attribute_id',
         'shopware6_id',
     ];
+
     /**
      * @var Connection
      */
@@ -38,43 +40,46 @@ class DbalShopware6MultimediaQuery implements Shopware6MultimediaQueryInterface
     }
 
     /**
-     * @param ChannelId $channel
+     * @param ChannelId $channelId
      * @param string    $shopwareId
      *
-     * @return MultimediaId|null
+     * @return AttributeId|null
      */
-    public function loadByShopwareId(ChannelId $channel, string $shopwareId): ?MultimediaId
+    public function loadByShopwareId(ChannelId $channelId, string $shopwareId): ?AttributeId
     {
         $query = $this->connection->createQueryBuilder();
         $record = $query
             ->select(self::FIELDS)
-            ->from(self::TABLE, 'm')
+            ->from(self::TABLE, 'pg')
             ->where($query->expr()->eq('channel_id', ':channelId'))
-            ->setParameter(':channelId', $channel->getValue())
+            ->setParameter(':channelId', $channelId->getValue())
             ->andWhere($query->expr()->eq('shopware6_id', ':shopware6Id'))
             ->setParameter(':shopware6Id', $shopwareId)
             ->execute()
             ->fetch();
 
         if ($record) {
-            return new MultimediaId($record['multimedia_id']);
+            return new AttributeId($record['attribute_id']);
         }
 
         return null;
     }
 
     /**
-     * @param ChannelId          $channel
+     * @param ChannelId          $channelId
      * @param \DateTimeImmutable $dateTime
+     * @param string             $type
      */
-    public function cleanData(ChannelId $channel, \DateTimeImmutable $dateTime): void
+    public function cleanData(ChannelId $channelId, \DateTimeImmutable $dateTime, string $type): void
     {
         $query = $this->connection->createQueryBuilder();
-        $query->delete(self::TABLE, 'm')
-            ->where($query->expr()->eq('m.channel_id', ':channelId'))
-            ->setParameter(':channelId', $channel->getValue())
-            ->andWhere($query->expr()->lt('m.update_at', ':updateAt'))
+        $query->delete(self::TABLE, 'pg')
+            ->where($query->expr()->eq('channel_id', ':channelId'))
+            ->setParameter(':channelId', $channelId->getValue())
+            ->andWhere($query->expr()->lt('pg.update_at', ':updateAt'))
             ->setParameter(':updateAt', $dateTime, Types::DATETIMETZ_MUTABLE)
+            ->andWhere($query->expr()->eq('pg.type', ':type'))
+            ->setParameter(':type', $type)
             ->execute();
     }
 }
