+++ module/exporter-shopware-6/src/Infrastructure/Persistence/Repository/DbalShopware6LanguageRepository.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ExporterShopware6\Persistence\Dbal\Repository;
+namespace Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Types\Types;
