+++ module/exporter-shopware-6/src/Infrastructure/Persistence/Repository/DbalShopware6PropertyGroupOptionsRepository.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -6,26 +7,27 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\ExporterShopware6\Persistence\Dbal\Repository;
+namespace Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository;
 
 use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\DBALException;
 use Doctrine\DBAL\Types\Types;
-use Ergonode\ExporterShopware6\Domain\Repository\Shopware6PropertyGroupRepositoryInterface;
+use Ergonode\ExporterShopware6\Domain\Repository\Shopware6PropertyGroupOptionsRepositoryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\SharedKernel\Domain\Aggregate\ChannelId;
-use Doctrine\DBAL\DBALException;
+use Ergonode\SharedKernel\Domain\AggregateId;
 
 /**
  */
-class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepositoryInterface
+class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGroupOptionsRepositoryInterface
 {
-    private const TABLE = 'exporter.shopware6_property_group';
+    private const TABLE = 'exporter.shopware6_property_group_options';
     private const FIELDS = [
         'channel_id',
         'attribute_id',
         'shopware6_id',
+        'option_id',
     ];
-
     /**
      * @var Connection
      */
@@ -42,19 +44,22 @@ class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepo
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
+     * @param AggregateId $optionId
      *
      * @return string|null
      */
-    public function load(ChannelId $channelId, AttributeId $attributeId): ?string
+    public function load(ChannelId $channelId, AttributeId $attributeId, AggregateId $optionId): ?string
     {
         $query = $this->connection->createQueryBuilder();
         $record = $query
             ->select(self::FIELDS)
-            ->from(self::TABLE, 'pg')
-            ->where($query->expr()->eq('channel_id', ':channelId'))
+            ->from(self::TABLE, 'pgo')
+            ->where($query->expr()->eq('pgo.channel_id', ':channelId'))
             ->setParameter(':channelId', $channelId->getValue())
-            ->andWhere($query->expr()->eq('pg.attribute_id', ':attributeId'))
+            ->andWhere($query->expr()->eq('pgo.attribute_id', ':attributeId'))
             ->setParameter(':attributeId', $attributeId->getValue())
+            ->andWhere($query->expr()->eq('pgo.option_id', ':optionId'))
+            ->setParameter(':optionId', $optionId->getValue())
             ->execute()
             ->fetch();
 
@@ -68,35 +73,43 @@ class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepo
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
+     * @param AggregateId $optionId
      * @param string      $shopwareId
-     * @param string      $type
      *
      * @throws DBALException
      */
-    public function save(ChannelId $channelId, AttributeId $attributeId, string $shopwareId, string $type): void
-    {
-        if ($this->exists($channelId, $attributeId)) {
-            $this->update($channelId, $attributeId, $shopwareId);
+    public function save(
+        ChannelId $channelId,
+        AttributeId $attributeId,
+        AggregateId $optionId,
+        string $shopwareId
+    ): void {
+        if ($this->exists($channelId, $attributeId, $optionId)) {
+            $this->update($channelId, $attributeId, $optionId, $shopwareId);
         } else {
-            $this->insert($channelId, $attributeId, $shopwareId, $type);
+            $this->insert($channelId, $attributeId, $optionId, $shopwareId);
         }
     }
 
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
+     * @param AggregateId $optionId
      *
      * @return bool
      */
-    public function exists(ChannelId $channelId, AttributeId $attributeId): bool
+    public function exists(ChannelId $channelId, AttributeId $attributeId, AggregateId $optionId): bool
     {
+
         $query = $this->connection->createQueryBuilder();
         $result = $query->select(1)
-            ->from(self::TABLE)
-            ->where($query->expr()->eq('channel_id', ':channelId'))
+            ->from(self::TABLE, 'pgo')
+            ->where($query->expr()->eq('pgo.channel_id', ':channelId'))
             ->setParameter(':channelId', $channelId->getValue())
-            ->andWhere($query->expr()->eq('attribute_id', ':attributeId'))
+            ->andWhere($query->expr()->eq('pgo.attribute_id', ':attributeId'))
             ->setParameter(':attributeId', $attributeId->getValue())
+            ->andWhere($query->expr()->eq('pgo.option_id', ':optionId'))
+            ->setParameter(':optionId', $optionId->getValue())
             ->execute()
             ->rowCount();
 
@@ -110,12 +123,17 @@ class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepo
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
+     * @param AggregateId $optionId
      * @param string      $shopwareId
      *
      * @throws DBALException
      */
-    private function update(ChannelId $channelId, AttributeId $attributeId, string $shopwareId): void
-    {
+    private function update(
+        ChannelId $channelId,
+        AttributeId $attributeId,
+        AggregateId $optionId,
+        string $shopwareId
+    ): void {
         $this->connection->update(
             self::TABLE,
             [
@@ -125,6 +143,7 @@ class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepo
             [
                 'attribute_id' => $attributeId->getValue(),
                 'channel_id' => $channelId->getValue(),
+                'option_id' => $optionId->getValue(),
             ],
             [
                 'update_at' => Types::DATETIMETZ_MUTABLE,
@@ -135,19 +154,23 @@ class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepo
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
+     * @param AggregateId $optionId
      * @param string      $shopwareId
-     * @param string      $type
      *
      * @throws DBALException
      */
-    private function insert(ChannelId $channelId, AttributeId $attributeId, string $shopwareId, string $type): void
-    {
+    private function insert(
+        ChannelId $channelId,
+        AttributeId $attributeId,
+        AggregateId $optionId,
+        string $shopwareId
+    ): void {
         $this->connection->insert(
             self::TABLE,
             [
                 'shopware6_id' => $shopwareId,
                 'attribute_id' => $attributeId->getValue(),
-                'type' => $type,
+                'option_id' => $optionId->getValue(),
                 'channel_id' => $channelId->getValue(),
                 'update_at' => new \DateTimeImmutable(),
             ],
