+++ module/exporter-shopware-6/src/Infrastructure/Persistence/Repository/DbalShopware6PropertyGroupRepository.php
@@ -1,32 +1,32 @@
 <?php
-/*
+
+/**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
  */
 
 declare(strict_types = 1);
 
-namespace Ergonode\ExporterShopware6\Persistence\Dbal\Repository;
+namespace Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DBALException;
 use Doctrine\DBAL\Types\Types;
-use Ergonode\ExporterShopware6\Domain\Repository\Shopware6PropertyGroupOptionsRepositoryInterface;
+use Ergonode\ExporterShopware6\Domain\Repository\Shopware6PropertyGroupRepositoryInterface;
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\SharedKernel\Domain\Aggregate\ChannelId;
-use Ergonode\SharedKernel\Domain\AggregateId;
 
 /**
  */
-class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGroupOptionsRepositoryInterface
+class DbalShopware6PropertyGroupRepository implements Shopware6PropertyGroupRepositoryInterface
 {
-    private const TABLE = 'exporter.shopware6_property_group_options';
+    private const TABLE = 'exporter.shopware6_property_group';
     private const FIELDS = [
         'channel_id',
         'attribute_id',
         'shopware6_id',
-        'option_id',
     ];
+
     /**
      * @var Connection
      */
@@ -43,22 +43,19 @@ class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGr
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
-     * @param AggregateId $optionId
      *
      * @return string|null
      */
-    public function load(ChannelId $channelId, AttributeId $attributeId, AggregateId $optionId): ?string
+    public function load(ChannelId $channelId, AttributeId $attributeId): ?string
     {
         $query = $this->connection->createQueryBuilder();
         $record = $query
             ->select(self::FIELDS)
-            ->from(self::TABLE, 'pgo')
-            ->where($query->expr()->eq('pgo.channel_id', ':channelId'))
+            ->from(self::TABLE, 'pg')
+            ->where($query->expr()->eq('channel_id', ':channelId'))
             ->setParameter(':channelId', $channelId->getValue())
-            ->andWhere($query->expr()->eq('pgo.attribute_id', ':attributeId'))
+            ->andWhere($query->expr()->eq('pg.attribute_id', ':attributeId'))
             ->setParameter(':attributeId', $attributeId->getValue())
-            ->andWhere($query->expr()->eq('pgo.option_id', ':optionId'))
-            ->setParameter(':optionId', $optionId->getValue())
             ->execute()
             ->fetch();
 
@@ -72,43 +69,35 @@ class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGr
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
-     * @param AggregateId $optionId
      * @param string      $shopwareId
+     * @param string      $type
      *
      * @throws DBALException
      */
-    public function save(
-        ChannelId $channelId,
-        AttributeId $attributeId,
-        AggregateId $optionId,
-        string $shopwareId
-    ): void {
-        if ($this->exists($channelId, $attributeId, $optionId)) {
-            $this->update($channelId, $attributeId, $optionId, $shopwareId);
+    public function save(ChannelId $channelId, AttributeId $attributeId, string $shopwareId, string $type): void
+    {
+        if ($this->exists($channelId, $attributeId)) {
+            $this->update($channelId, $attributeId, $shopwareId);
         } else {
-            $this->insert($channelId, $attributeId, $optionId, $shopwareId);
+            $this->insert($channelId, $attributeId, $shopwareId, $type);
         }
     }
 
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
-     * @param AggregateId $optionId
      *
      * @return bool
      */
-    public function exists(ChannelId $channelId, AttributeId $attributeId, AggregateId $optionId): bool
+    public function exists(ChannelId $channelId, AttributeId $attributeId): bool
     {
-
         $query = $this->connection->createQueryBuilder();
         $result = $query->select(1)
-            ->from(self::TABLE, 'pgo')
-            ->where($query->expr()->eq('pgo.channel_id', ':channelId'))
+            ->from(self::TABLE)
+            ->where($query->expr()->eq('channel_id', ':channelId'))
             ->setParameter(':channelId', $channelId->getValue())
-            ->andWhere($query->expr()->eq('pgo.attribute_id', ':attributeId'))
+            ->andWhere($query->expr()->eq('attribute_id', ':attributeId'))
             ->setParameter(':attributeId', $attributeId->getValue())
-            ->andWhere($query->expr()->eq('pgo.option_id', ':optionId'))
-            ->setParameter(':optionId', $optionId->getValue())
             ->execute()
             ->rowCount();
 
@@ -122,17 +111,12 @@ class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGr
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
-     * @param AggregateId $optionId
      * @param string      $shopwareId
      *
      * @throws DBALException
      */
-    private function update(
-        ChannelId $channelId,
-        AttributeId $attributeId,
-        AggregateId $optionId,
-        string $shopwareId
-    ): void {
+    private function update(ChannelId $channelId, AttributeId $attributeId, string $shopwareId): void
+    {
         $this->connection->update(
             self::TABLE,
             [
@@ -142,7 +126,6 @@ class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGr
             [
                 'attribute_id' => $attributeId->getValue(),
                 'channel_id' => $channelId->getValue(),
-                'option_id' => $optionId->getValue(),
             ],
             [
                 'update_at' => Types::DATETIMETZ_MUTABLE,
@@ -153,23 +136,19 @@ class DbalShopware6PropertyGroupOptionsRepository implements Shopware6PropertyGr
     /**
      * @param ChannelId   $channelId
      * @param AttributeId $attributeId
-     * @param AggregateId $optionId
      * @param string      $shopwareId
+     * @param string      $type
      *
      * @throws DBALException
      */
-    private function insert(
-        ChannelId $channelId,
-        AttributeId $attributeId,
-        AggregateId $optionId,
-        string $shopwareId
-    ): void {
+    private function insert(ChannelId $channelId, AttributeId $attributeId, string $shopwareId, string $type): void
+    {
         $this->connection->insert(
             self::TABLE,
             [
                 'shopware6_id' => $shopwareId,
                 'attribute_id' => $attributeId->getValue(),
-                'option_id' => $optionId->getValue(),
+                'type' => $type,
                 'channel_id' => $channelId->getValue(),
                 'update_at' => new \DateTimeImmutable(),
             ],
