+++ module/exporter-shopware-6/src/Resources/config/services.yml
@@ -14,9 +14,6 @@ services:
         resource: '../../Infrastructure/*'
         exclude: '../../Infrastructure/{Model}'
 
-    Ergonode\ExporterShopware6\Persistence\:
-        resource: '../../Persistence/*'
-
     Ergonode\ExporterShopware6\Infrastructure\Handler\:
         resource: '../../Infrastructure/Handler/*'
         tags: ['messenger.message_handler']
@@ -31,16 +28,16 @@ services:
                 - '@Ergonode\ExporterShopware6\Infrastructure\Processor\Step\Shopware6SimpleProductStep'
                 - '@Ergonode\ExporterShopware6\Infrastructure\Processor\Step\Shopware6CategoryRemoveStep'
     # Repository
-    Ergonode\ExporterShopware6\Domain\Repository\Shopware6CategoryRepositoryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Repository\DbalShopware6CategoryRepository'
-    Ergonode\ExporterShopware6\Domain\Repository\Shopware6CurrencyRepositoryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Repository\DbalShopware6CurrencyRepository'
-    Ergonode\ExporterShopware6\Domain\Repository\Shopware6TaxRepositoryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Repository\DbalShopware6TaxRepository'
-    Ergonode\ExporterShopware6\Domain\Repository\Shopware6PropertyGroupRepositoryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Repository\DbalShopware6PropertyGroupRepository'
-    Ergonode\ExporterShopware6\Domain\Repository\Shopware6CustomFieldRepositoryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Repository\DbalShopware6CustomFieldRepository'
+    Ergonode\ExporterShopware6\Domain\Repository\Shopware6CategoryRepositoryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository\DbalShopware6CategoryRepository'
+    Ergonode\ExporterShopware6\Domain\Repository\Shopware6CurrencyRepositoryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository\DbalShopware6CurrencyRepository'
+    Ergonode\ExporterShopware6\Domain\Repository\Shopware6TaxRepositoryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository\DbalShopware6TaxRepository'
+    Ergonode\ExporterShopware6\Domain\Repository\Shopware6PropertyGroupRepositoryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository\DbalShopware6PropertyGroupRepository'
+    Ergonode\ExporterShopware6\Domain\Repository\Shopware6CustomFieldRepositoryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Repository\DbalShopware6CustomFieldRepository'
 
     # Query
-    Ergonode\ExporterShopware6\Domain\Query\Shopware6CategoryQueryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Query\DbalShopware6CategoryQuery'
-    Ergonode\ExporterShopware6\Domain\Query\Shopware6PropertyGroupQueryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Query\DbalShopware6PropertyGroupQuery'
-    Ergonode\ExporterShopware6\Domain\Query\Shopware6CustomFieldQueryInterface: '@Ergonode\ExporterShopware6\Persistence\Dbal\Query\DbalShopware6CustomFiledQuery'
+    Ergonode\ExporterShopware6\Domain\Query\Shopware6CategoryQueryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Query\DbalShopware6CategoryQuery'
+    Ergonode\ExporterShopware6\Domain\Query\Shopware6PropertyGroupQueryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Query\DbalShopware6PropertyGroupQuery'
+    Ergonode\ExporterShopware6\Domain\Query\Shopware6CustomFieldQueryInterface: '@Ergonode\ExporterShopware6\Infrastructure\Persistence\Query\DbalShopware6CustomFiledQuery'
 
     # Synchronizer
     Ergonode\ExporterShopware6\Infrastructure\Processor\Process\StartShopware6ExportProcess:
